Сделал стрельбу Башен по врагам

+++ Assets/App/Scripts/Modifiers/Data/DamageModifierData.cs
@@ -7,7 +7,7 @@ namespace App.Scripts.Modifiers.Data
     [Serializable]
     public class DamageModifierData : BaseModifierData
     {
-        public float currentDamage;
+        public int currentDamage;
 
         public override void ResetToDefault(BaseModifierSO config)
         {
